user/requirements.txt
asgiref==3.7.2
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
grpcio==1.54.2
grpcio-tools==1.54.2
protobuf==4.23.2

